App/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
openai
python-dotenv
